Render SaveAndLoad component
Delete all ported code in App and use the new SaveAndLoad component
instead

Ensures better component composition, ensuring better readability of the
App component
